create controllers and tests
