+++ src/test/java/it/gov/pagopa/rtd/ms/rtdmsfilereporter/controller/FileReportControllerImplTest.java
@@ -8,12 +8,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import it.gov.pagopa.rtd.ms.rtdmsfilereporter.TestUtils;
-import it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.FileReportDto;
-import it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.FileReportDtoMapper;
-import it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.FileReportV2Dto;
-import it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.FileReportV2DtoMapper;
-import it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.v1.FileReportDto;
 import it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.v1.FileReportDtoMapper;
+import it.gov.pagopa.rtd.ms.rtdmsfilereporter.controller.model.v1.FileReportDto;
 import it.gov.pagopa.rtd.ms.rtdmsfilereporter.domain.model.FileReport;
 import it.gov.pagopa.rtd.ms.rtdmsfilereporter.domain.service.FileReportService;
 import java.util.Collections;
@@ -37,7 +33,6 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 class FileReportControllerImplTest {
 
   private final String FILE_REPORT_URL = "/file-report";
-  private final String FILE_REPORT_URL_V2 = "/v2/file-report";
   private final String SENDER_ADE_ACK_URL = "/sender-ade-ack";
 
   ObjectMapper objectMapper = new ObjectMapper();
@@ -47,9 +42,6 @@ class FileReportControllerImplTest {
 
   @MockBean
   private FileReportDtoMapper mapper;
-
-  @MockBean
-  private FileReportV2DtoMapper mapperV2;
 
   @MockBean
   private FileReportService fileReportService;
@@ -87,24 +79,6 @@ class FileReportControllerImplTest {
         .andReturn();
     FileReportDto fileReportResponse = objectMapper.readValue(
         result.getResponse().getContentAsString(), FileReportDto.class);
-
-    assertThat(fileReportResponse.getFilesRecentlyUploaded()).isNotNull().hasSize(2);
-  }
-
-  @SneakyThrows
-  @Test
-  void givenReportWhenGetFileReportV2ThenReturnCorrectJson() {
-    var reportMock = TestUtils.createFileReport(2, 2);
-    var reportDto = Mappers.getMapper(FileReportV2DtoMapper.class).fileReportToDto(reportMock);
-    Mockito.when(mapperV2.fileReportToDto(any())).thenReturn(reportDto);
-
-    MvcResult result = mockMvc.perform(
-            MockMvcRequestBuilders.get(FILE_REPORT_URL_V2).param("senderCodes", "12345"))
-        .andExpectAll(status().isOk(),
-            content().contentType(MediaType.APPLICATION_JSON_VALUE))
-        .andReturn();
-    FileReportV2Dto fileReportResponse = objectMapper.readValue(
-        result.getResponse().getContentAsString(), FileReportV2Dto.class);
 
     assertThat(fileReportResponse.getFilesRecentlyUploaded()).isNotNull().hasSize(2);
   }
